Fix and simplify Up&Down_Time constraints
The min-uptime and min-downtime constraints can be greatly simplified
using a bit of modular arithmetic. This also corrects some off-by-one
errors in the previous implementation, and changes the definition of
Up_Time to be <previous Up_Time> - 1.

Previously Up_Time = 1 meant a generator would need to be on for *two*
hours (except for some edge case bugs that resulted in either 1 or 3
hours), now it's uniformly one hour as expected.

Co-authored-by: Qingyu Xu <xuqingyu0610@gmail.com>