Functionalize Dataframe loading from csv
Add a function to load a dataframe from our common sort of CSV file.
This helps DRY with regard to dataframe loading settings, making code
easier to read and ensuring the same options are applied everywhere.
Also, if we want to make dataframe loading smarter, now there's one
centralized place to do it.

This commit adds a feature that CSV files of different case can also be
loaded. It does not look for folders or yaml files with different cases.