Unset executable bits (#420)
GenX does not actually have any binaries or scripts --- files starting with `#!/usr/bin/julia` or similar --- so having the executable permission bit set is not actually necessary. The only executable that GenX uses currently is the user's installation of `julia`.

This commit unsets the executable bit for many GenX files.