Fixed issue #461 and bug in policy constraint slacks (#464)
This fixes an issue where constraints on @expressions for policies were defined before the expressions were fully-developed (with all terms added). This caused those terms to be ignored in the constraints.
Moving the constraint definitions to lines after the expressions are developed fixes the issue.

Separately, the Max_Cap_Requirement slack variables had an incorrect sign.